+++ NEWS.md
@@ -2,12 +2,17 @@
 
 ## Bug Fixes
 
- * `progressify()` for `mapply()`, `Map()`, `.mapply()`, and `apply()`
-   would produce an error on `object '.progressr_progressor' not found`.
+ * `progressify()` for `mapply()`, `Map()`, `.mapply()`, `apply()`, and
+   `tapply()` would produce an error on
+   `object '.progressr_progressor' not found`.
 
  * `progressify()` for `future_mapply()`, `future_Map()`, and
    `future_.mapply()` of **future.apply** would produce an error on
    `length(idx_FUN) == 1L is not TRUE`.
+
+ * `progressify()` for `future_apply()` and `future_tapply()` of
+   **future.apply** would produce an error on
+   `object '.progressr_progressor' not found`.
 
 
 # Version 0.2.0 [2026-05-31]